refactor: 패키지 구조 변경, facade 적용
