+++ src/main/java/com/abab/lectureRegister/LectureBookingServiceApplication.java
@@ -1,4 +1,4 @@
-package com.abab.lecture;
+package com.abab.lectureRegister;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
