+++ src/main/java/com/abab/lectureRegister/lecture/LectureRepository.java
@@ -1,0 +1,11 @@
+package com.abab.lectureRegister.lecture;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public interface LectureRepository extends JpaRepository<Lecture, Long> {
+    List<Lecture> findAllByStartTimeAndCurrentEnrollmentLessThan(LocalDateTime startTime, int currentEnrollment);
+    boolean existsByLectureIdAndCurrentEnrollmentGreaterThanEqual(Long lectureId, int currentEnrollment);
+}
