+++ src/main/java/com/abab/lectureRegister/lecture/LectureService.java
@@ -1,0 +1,12 @@
+package com.abab.lectureRegister.lecture;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LectureService {
+    private final LectureRepository lectureRepository;
+
+
+}
