+++ src/main/java/com/abab/lectureRegister/registration/RegistrationService.java
@@ -1,0 +1,11 @@
+package com.abab.lectureRegister.registration;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class RegistrationService {
+    private final RegistrationRepository registrationRepository;
+
+}
