+++ src/test/java/com/abab/lectureRegister/LectureBookingServiceApplicationTests.java
@@ -1,4 +1,4 @@
-package com.abab.lecture;
+package com.abab.lectureRegister;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
